chore: update icons
