+++ packages/ui/src/components/icons/index.ts
@@ -40,9 +40,9 @@ export { default as DropdownDownIcon } from "./DropdownDownIcon"
 export { default as DropdownUpIcon } from "./DropdownUpIcon"
 export { default as DeployIcon } from "./DeployIcon"
 export { default as ApproveIcon } from "./ApproveIcon"
-export { default as DesktopIcon } from "./DesktopIcon"
-export { default as WebIcon } from "./WebIcon"
 export { default as EmailIcon } from "./EmailIcon"
 export { default as ProfileIcon } from "./ProfileIcon"
 export { default as InfoIcon } from "./InfoIcon"
 export { default as TickIcon } from "./TickIcon"
+export { default as LinkIcon } from "./LinkIcon"
+export { default as EditIcon } from "./EditIcon"
